feat(theme): removes themes parsing
