--- assets/themes.toml
@@ -1,20 +1,0 @@
-default = "Amber"
-
-[[themes]]
-name = "Amber"
-mode = "dark"
-[themes.colors.accent_color]
-l = 0.6388
-chroma = 0.1435
-hue = 64.8
-alpha = 98
-[themes.colors.background_color]
-l = 0.1638
-chroma = 0.0
-hue = 0.0
-alpha = 100.0
-[themes.colors.foreground_color]
-l = 0.8638
-chroma = 0.0
-hue = 0.0
-alpha = 100.0
