+++ shell/crates/theme/src/colors.rs
@@ -1,4 +1,6 @@
+use crate::prelude::*;
 use gpui::*;
+use palette::{Darken, IntoColor, Lighten, Oklcha};
 
 #[derive(Debug, Clone, Default)]
 pub struct ThemeColors {
@@ -41,4 +43,89 @@ pub struct ThemeColors {
     pub foreground_1000: Rgba,
     pub foreground_1100: Rgba,
     pub foreground_1200: Rgba,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ColorsSetting {
+    pub accent_color: Oklcha,
+    pub background_color: Oklcha,
+    pub foreground_color: Oklcha,
+}
+
+impl Default for ColorsSetting {
+    fn default() -> Self {
+        Self {
+            accent_color: Oklcha::new(0.6388, 0.1435, 64.8, 98.0),
+            background_color: Oklcha::new(0.1638, 0.0, 0.0, 100.0),
+            foreground_color: Oklcha::new(0.8638, 0.0, 0.0, 100.0),
+        }
+    }
+}
+
+impl ThemeColors {
+    pub fn apply_setting(&mut self, mode: ThemeMode, colors: ColorsSetting) {
+        let ColorsSetting {
+            accent_color,
+            foreground_color,
+            background_color,
+            ..
+        } = colors;
+
+        println!(
+            "Setting colors: {:#?} {:#?} {:#?}",
+            accent_color, foreground_color, background_color
+        );
+
+        self.accent_0 = oklcha_to_rgba(accent_color.lighten_fixed(0.6));
+        self.accent_100 = oklcha_to_rgba(accent_color.lighten_fixed(0.5));
+        self.accent_200 = oklcha_to_rgba(accent_color.lighten_fixed(0.4));
+        self.accent_300 = oklcha_to_rgba(accent_color.lighten_fixed(0.3));
+        self.accent_400 = oklcha_to_rgba(accent_color.lighten_fixed(0.2));
+        self.accent_500 = oklcha_to_rgba(accent_color.lighten_fixed(0.1));
+        self.accent_600 = oklcha_to_rgba(accent_color);
+        self.accent_700 = oklcha_to_rgba(accent_color.darken_fixed(0.1));
+        self.accent_800 = oklcha_to_rgba(accent_color.darken_fixed(0.2));
+        self.accent_900 = oklcha_to_rgba(accent_color.darken_fixed(0.3));
+        self.accent_1000 = oklcha_to_rgba(accent_color.darken_fixed(0.4));
+        self.accent_1100 = oklcha_to_rgba(accent_color.darken_fixed(0.5));
+        self.accent_1200 = oklcha_to_rgba(accent_color.darken_fixed(0.6));
+
+        self.background_0 = oklcha_to_rgba(background_color.lighten_fixed(0.6));
+        self.background_100 = oklcha_to_rgba(background_color.lighten_fixed(0.5));
+        self.background_200 = oklcha_to_rgba(background_color.lighten_fixed(0.4));
+        self.background_300 = oklcha_to_rgba(background_color.lighten_fixed(0.3));
+        self.background_400 = oklcha_to_rgba(background_color.lighten_fixed(0.2));
+        self.background_500 = oklcha_to_rgba(background_color.lighten_fixed(0.1));
+        self.background_600 = oklcha_to_rgba(background_color);
+        self.background_700 = oklcha_to_rgba(background_color.darken_fixed(0.1));
+        self.background_800 = oklcha_to_rgba(background_color.darken_fixed(0.2));
+        self.background_900 = oklcha_to_rgba(background_color.darken_fixed(0.3));
+        self.background_1000 = oklcha_to_rgba(background_color.darken_fixed(0.4));
+        self.background_1100 = oklcha_to_rgba(background_color.darken_fixed(0.5));
+        self.background_1200 = oklcha_to_rgba(background_color.darken_fixed(0.6));
+
+        self.foreground_0 = oklcha_to_rgba(foreground_color.lighten_fixed(0.6));
+        self.foreground_100 = oklcha_to_rgba(foreground_color.lighten_fixed(0.5));
+        self.foreground_200 = oklcha_to_rgba(foreground_color.lighten_fixed(0.4));
+        self.foreground_300 = oklcha_to_rgba(foreground_color.lighten_fixed(0.3));
+        self.foreground_400 = oklcha_to_rgba(foreground_color.lighten_fixed(0.2));
+        self.foreground_500 = oklcha_to_rgba(foreground_color.lighten_fixed(0.1));
+        self.foreground_600 = oklcha_to_rgba(foreground_color);
+        self.foreground_700 = oklcha_to_rgba(foreground_color.darken_fixed(0.1));
+        self.foreground_800 = oklcha_to_rgba(foreground_color.darken_fixed(0.2));
+        self.foreground_900 = oklcha_to_rgba(foreground_color.darken_fixed(0.3));
+        self.foreground_1000 = oklcha_to_rgba(foreground_color.darken_fixed(0.4));
+        self.foreground_1100 = oklcha_to_rgba(foreground_color.darken_fixed(0.5));
+        self.foreground_1200 = oklcha_to_rgba(foreground_color.darken_fixed(0.6));
+    }
+}
+
+pub fn oklcha_to_rgba(color: Oklcha) -> Rgba {
+    let rgba: palette::rgb::Rgba = color.into_color();
+    Rgba {
+        r: rgba.red,
+        g: rgba.green,
+        b: rgba.blue,
+        a: rgba.alpha,
+    }
 }
